fix: refactore
