+++ src/app/components/sub/BurgerButton.jsx
@@ -1,0 +1,15 @@
+import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline';
+
+const BurgerButton = ({ toggleMenu, navBarIsOpen }) => {
+  return (
+    <button className="lg:hidden" onClick={toggleMenu}>
+      {navBarIsOpen ? (
+        <XMarkIcon className="h-6 w-6 text-darkBlue" />
+      ) : (
+        <Bars3Icon className="h-6 w-6 text-darkBlue" />
+      )}
+    </button>
+  );
+};
+
+export default BurgerButton;
